Move compute_edge_combined_degree
