+++ graphrag/index/verbs/graph/__init__.py
@@ -4,7 +4,6 @@
 """The Indexing Engine graph package root."""
 
 from .clustering import cluster_graph
-from .compute_edge_combined_degree import compute_edge_combined_degree
 from .create import DEFAULT_EDGE_ATTRIBUTES, DEFAULT_NODE_ATTRIBUTES, create_graph
 from .embed import embed_graph
 from .layout import layout_graph
@@ -21,7 +20,6 @@ __all__ = [
     "DEFAULT_EDGE_ATTRIBUTES",
     "DEFAULT_NODE_ATTRIBUTES",
     "cluster_graph",
-    "compute_edge_combined_degree",
     "create_community_reports",
     "create_graph",
     "embed_graph",
